fix utils/decorators test
