+++ tests/test_decorators.py
@@ -1,5 +1,3 @@
-from tempfile import TemporaryDirectory
-from pathlib import Path
 
 import click
 from click.testing import CliRunner
@@ -11,7 +9,6 @@ from ocrd.decorators import (
     ocrd_cli_options,
     ocrd_loglevel,
     ocrd_cli_wrap_processor,
-    _parse_json_string_or_file
 )    # pylint: disable=protected-access
 from ocrd_utils.logging import setOverrideLogLevel, initLogging
 from ocrd_utils import pushd_popd
@@ -88,27 +85,6 @@ class TestDecorators(TestCase):
             with pushd_popd(tempdir):
                 result = self.runner.invoke(cli_dummy_processor, ['--mets', 'mets.xml'])
                 self.assertEqual(result.exit_code, 0)
-
-    def test_parameters0(self):
-        self.assertEqual(_parse_json_string_or_file(None, None), {})
-        self.assertEqual(_parse_json_string_or_file(None, None, '{}'), {})
-        self.assertEqual(_parse_json_string_or_file(None, None, '{"foo": 32}'), {'foo': 32})
-        self.assertEqual(_parse_json_string_or_file(None, None, '{"foo": 32}'), {'foo': 32})
-
-    def test_parameter_file(self):
-        with TemporaryDirectory() as tempdir:
-            paramfile = str(Path(tempdir, '{}')) # XXX yes, the file is called '{}'
-            with open(paramfile, 'w') as f:
-                f.write('{"bar": 42}')
-            self.assertEqual(_parse_json_string_or_file(None, None, paramfile), {'bar': 42})
-            with pushd_popd(tempdir):
-                self.assertEqual(_parse_json_string_or_file(None, None), {'bar': 42})
-
-    def test_parameters_invalid(self):
-        with self.assertRaisesRegex(ValueError, 'Not a valid JSON object'):
-            _parse_json_string_or_file(None, None, '[]')
-        with self.assertRaisesRegex(ValueError, 'Error parsing'):
-            _parse_json_string_or_file(None, None, '[}')
 
 
 if __name__ == '__main__':
